SCRUM-179, SCRUM-188: モノレポプロジェクトにリファクタ＆本番用Dockerfileの作成
- Mavenにアップ済み
