+++ .env.example
@@ -1,0 +1,98 @@
+# =============================================================================
+# Kishax API Multi-Module Environment Configuration
+# =============================================================================
+# Unified environment file for both local and production environments
+# Copy this file to .env and configure as needed
+# Set ENVIRONMENT=local or ENVIRONMENT=production to switch configurations
+
+# =============================================================================
+# Environment Mode (local/production)
+# =============================================================================
+ENVIRONMENT=local
+
+# =============================================================================
+# Integration Test Control
+# =============================================================================
+# LocalStack tests (enabled for local, disabled for production)
+RUN_INTEGRATION_TESTS=true
+# Real AWS tests (disabled for local, enabled for production)
+RUN_REAL_AWS_TESTS=false
+
+# =============================================================================
+# AWS SQS Configuration
+# =============================================================================
+AWS_REGION=ap-northeast-1
+
+# Local (LocalStack): use test credentials
+# Production: use your actual AWS credentials
+MC_WEB_SQS_ACCESS_KEY_ID=test
+MC_WEB_SQS_SECRET_ACCESS_KEY=test
+
+# Local (LocalStack): use localhost endpoints
+# Production: use actual SQS queue URLs
+MC_TO_WEB_QUEUE_URL=http://localhost:4566/000000000000/mc-web-queue
+WEB_TO_MC_QUEUE_URL=http://localhost:4566/000000000000/web-mc-queue
+
+# =============================================================================
+# Redis Configuration
+# =============================================================================
+# Local: redis://localhost:6379 (different port from other services)
+# Production: redis://your-redis-cluster.cache.amazonaws.com:6379
+# you can overwrite REDIS_URL at compose.yaml
+# REDIS_URL=redis://localhost:6379
+REDIS_CONNECTION_TIMEOUT=5000
+REDIS_COMMAND_TIMEOUT=3000
+
+# =============================================================================
+# SQS Worker Configuration
+# =============================================================================
+QUEUE_MODE=MC
+SQS_WORKER_ENABLED=true
+SQS_WORKER_POLLING_INTERVAL_SECONDS=5
+SQS_WORKER_MAX_MESSAGES=10
+SQS_WORKER_WAIT_TIME_SECONDS=20
+SQS_WORKER_VISIBILITY_TIMEOUT_SECONDS=30
+
+# =============================================================================
+# Authentication API Configuration
+# =============================================================================
+AUTH_API_ENABLED=true
+AUTH_API_PORT=8080
+# Local: local_dev_auth_key
+# Production: your_production_auth_key_here
+AUTH_API_KEY=local_dev_auth_key
+
+# =============================================================================
+# Database Configuration
+# =============================================================================
+# Local: postgresql://localhost:5432/kishax_dev?user=dev_user&password=dev_pass
+# Production: postgresql://your-rds-endpoint.region.rds.amazonaws.com:5432/kishax_prod?user=prod_user&password=prod_pass
+# see default kishax-web's postgres service info at https://github.com/Kishax/kishax-web/blob/master/compose.yaml
+# you can overwrite DATABASE_URL at compose.yaml
+# DATABASE_URL=jdbc:postgresql://localhost:5432/kishax_db?user=postgres&password=postgres
+
+# =============================================================================
+# Web API Configuration
+# =============================================================================
+# Local: http://localhost:3000
+# Production: http://web-service.kishax.internal:3000 (ECS internal networking)
+WEB_API_URL=http://localhost:3000
+# Local: local_dev_api_key
+# Production: your_production_web_api_key_here
+WEB_API_KEY=local_dev_api_key
+
+# =============================================================================
+# Application Configuration
+# =============================================================================
+# Local: 10 seconds
+# Production: 30 seconds
+APP_SHUTDOWN_GRACE_PERIOD_SECONDS=10
+
+# =============================================================================
+# Logging Configuration
+# =============================================================================
+# Local: DEBUG
+# Production: INFO
+LOG_LEVEL=DEBUG
+AWS_SDK_LOG_LEVEL=WARN
+LETTUCE_LOG_LEVEL=WARN
