--- .env.local.example
@@ -1,37 +1,0 @@
-# =============================================================================
-# LocalStack Development Environment
-# =============================================================================
-# Settings for local development testing with Docker Compose + LocalStack
-
-# Enable LocalStack integration tests
-RUN_INTEGRATION_TESTS=true
-
-# Disable real AWS tests
-RUN_REAL_AWS_TESTS=false
-
-# =============================================================================
-# LocalStack Configuration (Automatically set by docker compose)
-# =============================================================================
-# These values are used when running with docker compose up
-# No need to modify these for local development
-
-# Redis (local)
-REDIS_URL=redis://localhost:6380
-
-# SQS Configuration for LocalStack
-AWS_REGION=us-east-1
-AWS_SQS_ACCESSKEYID=test
-AWS_SQS_SECRETACCESSKEY=test
-AWS_SQS_MCTOWEBQUEUEURL=http://localhost:4566/000000000000/mc-web-queue
-AWS_SQS_WEBTOMCQUEUEURL=http://localhost:4566/000000000000/web-mc-queue
-
-# Logging
-LOG_LEVEL=DEBUG
-AWS_SDK_LOG_LEVEL=WARN
-LETTUCE_LOG_LEVEL=WARN
-
-# Web
-WEB_API_URL=http://localhost:3000
-WEB_API_KEY=local_dev_api_key
-
-KISHAX_AWS_VERSION=1.0.4
