--- .env.prod.example
@@ -1,41 +1,0 @@
-# =============================================================================
-# Production AWS Environment
-# =============================================================================
-# Settings for testing with real AWS services
-# IMPORTANT: Fill in your actual AWS credentials and queue URLs
-
-# Enable real AWS tests
-RUN_REAL_AWS_TESTS=true
-
-# Disable LocalStack tests
-RUN_INTEGRATION_TESTS=false
-
-# =============================================================================
-# AWS Configuration (REQUIRED - Fill these in!)
-# =============================================================================
-AWS_REGION=us-east-1
-AWS_SQS_ACCESSKEYID=your_actual_access_key_here
-AWS_SQS_SECRETACCESSKEY=your_actual_secret_key_here
-
-# SQS Queue URLs (Create these in your AWS account)
-AWS_SQS_MCTOWEBQUEUEURL=https://sqs.us-east-1.amazonaws.com/123456789012/mc-web-queue
-AWS_SQS_WEBTOMCQUEUEURL=https://sqs.us-east-1.amazonaws.com/123456789012/web-mc-queue
-
-# =============================================================================
-# Redis Configuration (Production)
-# =============================================================================
-# For production Redis (AWS ElastiCache, etc.)
-REDIS_URL=redis://your-redis-cluster.cache.amazonaws.com:6379
-
-# =============================================================================
-# Logging (Production)
-# =============================================================================
-LOG_LEVEL=INFO
-AWS_SDK_LOG_LEVEL=WARN
-LETTUCE_LOG_LEVEL=WARN
-
-# Web
-WEB_API_URL=http://localhost:3000
-WEB_API_KEY=local_dev_api_key
-
-KISHAX_AWS_VERSION=1.0.4
