+++ common/src/main/resources/application.properties
@@ -1,0 +1,35 @@
+# Kishax AWS Integration Configuration
+
+# AWS Configuration
+aws.region=${AWS_REGION:ap-northeast-1}
+aws.sqs.accessKeyId=${MC_WEB_SQS_ACCESS_KEY_ID:}
+aws.sqs.secretAccessKey=${MC_WEB_SQS_SECRET_ACCESS_KEY:}
+aws.sqs.mcToWebQueueUrl=${MC_TO_WEB_QUEUE_URL:}
+aws.sqs.webToMcQueueUrl=${WEB_TO_MC_QUEUE_URL:}
+
+# Redis Configuration
+redis.url=${REDIS_URL:redis://localhost:6379}
+redis.connectionTimeout=${REDIS_CONNECTION_TIMEOUT:5000}
+redis.commandTimeout=${REDIS_COMMAND_TIMEOUT:3000}
+
+# SQS Worker Configuration
+sqs.worker.enabled=${SQS_WORKER_ENABLED:false}
+sqs.worker.pollingIntervalSeconds=${SQS_WORKER_POLLING_INTERVAL:5}
+sqs.worker.maxMessages=${SQS_WORKER_MAX_MESSAGES:10}
+sqs.worker.waitTimeSeconds=${SQS_WORKER_WAIT_TIME:20}
+sqs.worker.visibilityTimeoutSeconds=${SQS_WORKER_VISIBILITY_TIMEOUT:300}
+
+# WEB API Configuration
+web.api.url=${WEB_API_URL:http://localhost:3000}
+web.api.key=${WEB_API_KEY:}
+
+# Application Configuration
+app.shutdownGracePeriodSeconds=${SHUTDOWN_GRACE_PERIOD:10}
+
+# Auth API Configuration
+auth.api.enabled=${AUTH_API_ENABLED:false}
+auth.api.port=${AUTH_API_PORT:8080}
+auth.api.key=${AUTH_API_KEY:local-dev-api-key}
+
+# Database Configuration
+database.url=${DATABASE_URL:jdbc:postgresql://localhost:5432/kishax}
