+++ mc-auth/src/main/java/net/kishax/api/auth/AuthApiController.java
