+++ mc-auth/src/main/java/net/kishax/api/auth/AuthCheckRequest.java
