+++ mc-auth/src/main/java/net/kishax/api/auth/AuthLevelResponse.java
