+++ mc-auth/src/test/java/net/kishax/api/auth/AuthApiIntegrationTest.java
