--- src/main/resources/application.properties
@@ -1,34 +1,0 @@
-# Kishax AWS Integration Configuration
-
-# AWS Configuration
-aws.region=${AWS_REGION:ap-northeast-1}
-aws.sqs.accessKeyId=${MC_WEB_SQS_ACCESS_KEY_ID:}
-aws.sqs.secretAccessKey=${MC_WEB_SQS_SECRET_ACCESS_KEY:}
-aws.sqs.mcToWebQueueUrl=${MC_TO_WEB_QUEUE_URL:}
-aws.sqs.webToMcQueueUrl=${WEB_TO_MC_QUEUE_URL:}
-
-# Redis Configuration
-redis.url=${REDIS_URL:redis://localhost:6379}
-redis.connectionTimeout=${REDIS_CONNECTION_TIMEOUT:5000}
-redis.commandTimeout=${REDIS_COMMAND_TIMEOUT:3000}
-
-# SQS Worker Configuration
-sqs.worker.enabled=${SQS_WORKER_ENABLED:true}
-sqs.worker.pollingIntervalSeconds=${SQS_WORKER_POLLING_INTERVAL:5}
-sqs.worker.maxMessages=${SQS_WORKER_MAX_MESSAGES:10}
-sqs.worker.waitTimeSeconds=${SQS_WORKER_WAIT_TIME:20}
-sqs.worker.visibilityTimeoutSeconds=${SQS_WORKER_VISIBILITY_TIMEOUT:300}
-
-# Logging Configuration
-logging.level.net.kishax.aws=${LOG_LEVEL:INFO}
-logging.level.software.amazon.awssdk=${AWS_SDK_LOG_LEVEL:WARN}
-logging.level.io.lettuce=${LETTUCE_LOG_LEVEL:WARN}
-
-# WEB API Configuration
-web.api.url=${WEB_API_URL:http://localhost:3000}
-web.api.key=${WEB_API_KEY:}
-
-# Application Configuration
-app.name=Kishax AWS Integration
-app.version=1.0.4
-app.shutdownGracePeriodSeconds=${SHUTDOWN_GRACE_PERIOD:10}
